이전 커밋 버그 해결
Object class의 indices의 선언이 byte가 아니라 int type으로 되어있었음(...)
- 사소한것도 잘 확인하자

+++ CG_proj_03/Prob_16/shaders.h
@@ -64,7 +64,7 @@ protected:
 
     vector<float> vertices;
     vector<float> colors;
-    vector<GLuint> indices;
+    vector<GLubyte> indices;
 
     glm::mat4 transformMat;
